consensus: reshape the hashing blob for stock XMRig (4-byte nonce at offset 39, ExtraNonce in the seal) (#52)
* consensus: reshape the hashing blob for stock XMRig

Header.PoWInput was 32 bytes of seed followed by a little-endian 64-bit
nonce: 40 bytes, nonce at offset 32. Stock XMRig writes a four-byte
little-endian nonce at offset 39 of whatever blob it is handed for rx/0,
because that is where Monero's block-hashing blob puts it, and the offset
is compiled in. A chain that wants unmodified XMRig pointed at it does not
get to choose another number.

The blob is now 43 bytes:

    0..31   PoWSeed()      blake3(TagPoW, SSZ(header with Nonce zeroed))
    32..38  seven zeroes   reserved, MUST be zero
    39..42  le32(Nonce)    the searched nonce

PoWSeal splits its 64-bit Nonce into Nonce and ExtraNonce, both 32 bits,
so the seal is still PoWSealSize bytes and HeaderSize does not move.
PoWSeed zeroes Nonce and only Nonce, so ExtraNonce stays inside the seed
preimage: every value of it is a distinct seed and therefore a disjoint
2^32 nonce space, which is how a pool shards work across stock miners.
Monero pools get the same separation by varying tx_extra in the coinbase
transaction; this chain has no coinbase transaction to vary. Solo mining
sets it to 0.

pow.Solver builds its buffer from the named offsets rather than from "the
last eight bytes", and Try takes a uint32 so that the narrowing is a
compile error rather than a silent truncation. pow.Solve and the miner's
attempt budget are clamped to the nonce space.

Test churn is mechanical: nonce salts that used bits above 32 move down
into the space that now exists. Two of them needed more than a shift —
ingressleak's per-caller windows would have overlapped above salt 255,
and the solver-agreement worker regions would have collapsed onto one —
and both are commented where they sit.

Refs #35.

* consensus: pin the blob layout and document the zero-pad rule

The layout tests are written against literal offsets rather than against
the constants they pin, because a test that reads PoWInputNonceOffset to
decide where to look for the nonce passes when somebody edits
PoWInputNonceOffset. One test is allowed to read the constants, and its
only job is to say that they describe the bytes.

Each of the new tests was checked against the mutation it exists to
catch: zeroing ExtraNonce alongside Nonce in the seed, writing the nonce
back at offset 32, and swapping the seal's two 32-bit halves on decode.
All three fail the suite and pass without it.

The zero-pad rule gets a normative paragraph in ARCHITECTURE.md because
it binds implementations rather than blocks: no header field can reach
bytes 32..38, so no block can violate it and no fold rule could reject
one. An implementation that filled the gap differently would fork on its
first block and no corpus of blocks could catch it.

docs/adversarial/I1.md gets a forward-pointer rather than a rewrite. Its
I1-M5 entry quotes the old pow_seed ‖ le64(nonce) form, but that file is
the audit's record of what was found and decided at the time, not a
second copy of the specification. The finding itself still holds: the
nonce still enters the input, and only the shape of the concatenation
moved.

Refs #35.

* docs: state the deferred work instead of pointing at issue numbers

This repository is published as files: a fresh git init under an
unlinked origin, with no history and no issue tracker. In that tree a
bare issue number resolves to nothing, and a sentence whose argument was
carried by one has quietly lost that argument — worse than a broken
link, because nothing looks broken. sim/wiring's history guard refuses
them, and five of mine were in the swept part of the tree.

Rewritten rather than deleted, which is what the guard's own header asks
for: deleting the token drops what it was carrying.

Three sites pointed at the missing blob vector. They now say what is
missing and why a block corpus cannot supply it — the vector would fix
the 43 bytes and the digest they produce against stock XMRig, and the
corpus in spec/vectors folds blocks, which cannot express a gap that no
header field can reach. Each names docs/ARCHITECTURE.md §12 as the
normative statement holding the rule in the meantime.

One pointed at the miner rewrite. It now states the fact a reader of
this tree needs: this miner never sets or rolls ExtraNonce, it mines at
zero, and the policy question left open is whether a solo miner reaching
the end of a space should prefer a fresh ExtraNonce or a fresh template.

One cited this change's own issue from the adversarial record. It now
cites the section that carries the layout, and states in-line the reason
the offset is not ours to choose.

* pow: make the solver/rule agreement test bite

TestTheSolverAgreesWithCheckWork is the only thing holding the property
that Try and CheckWork answer alike, and its own comment says the two
sharing no code is the hazard it exists to cover. It was not covering
it.

It ran at GenesisTarget alone, where none of its twelve boundary nonces
solves. So `viaRule` and `viaSolver` were both false at every nonce and
the assertion compared false against false. Measured rather than
suspected: with the test as it stood, a Try that wrote the nonce at the
old offset 32, and a Try that read the digest big-endian while
CheckWork read it little-endian, both passed it. Either defect ships a
miner whose blocks its own network rejects, and every test in the tree
that mines a block would still have been green, because they all mine
through this function.

It now runs at MaxTarget as well, where roughly one nonce in thirty-four
solves, so the true arm is exercised and a divergence has somewhere to
show. The twelve boundary values stay — they are what catch a wrong
offset or a wrong width — with a contiguous run of 256 added on top,
because the boundaries alone reach no solution at either target.

An anti-vacuity check states the requirement rather than leaving it to
luck: if no nonce solves at MaxTarget the test fails outright, so a
future engine or parameter change that moved the solve rate cannot
silently hollow this out again the way it was hollow before.

Both mutations above now fail it, and the pre-existing weakness is
noted rather than glossed: this vacuity predates the blob change and is
inherited, not introduced by it.

* test: confirm forged work against the rule's own digest convention

Five test-side comparisons still read the work-function digest
big-endian while the rule they are confirming reads it little-endian.
The two conventions read independent ends of the digest, so they agree
only by coincidence — and TestABlockCitingAHeaderThatNeverDidItsWorkIsRefused
is where the coincidence ran out: its setup loop finds a nonce by asking
CheckWork, then the very next line re-states the same fact through
FromBytes and reaches the opposite answer, failing with "the forged
citation meets its declared target after all".

The inconsistency is not new and is not the blob change's. It is present
verbatim on dev, where it passes because the first nonce tried happens
to satisfy both readings. Reshaping the blob moves every digest, the
luck does not survive the move, and the latent disagreement becomes a
failure. Left alone it would have gone on being a coin flip that
re-lands on every change to the hashing input.

Each site now uses FromLEBytes, so a confirmation asserts what the rule
actually says rather than something adjacent to it. The commented-out
illustration of the rule in flood_internal_test.go is corrected for the
same reason: a comment that quotes the check wrongly teaches the wrong
check.

+++ node/p2p/keyepoch_internal_test.go
@@ -134,7 +134,7 @@ func TestOneBlockBodyCannotBuyManyKeyEpochs(t *testing.T) {
 			ParentID: types.Hash{byte(i)},
 			Time:     p.GenesisTime + h*p.TargetBlockSeconds,
 			Target:   u256.Max,
-			PoW:      types.PoWSeal{Nonce: i},
+			PoW:      types.PoWSeal{Nonce: uint32(i)},
 		})
 	}
 	blk.Header.CertRoot = blk.ComputeCertRoot(p)
